[feat][patch] 편집 가능한 테이블 형성 및 제공, 버전 삭제 캐밥, 크론 생성 캐밥 생성, app 페이지 리팩토링
