+++ frontend/public/components/hypercloud/sas/sas-app.tsx
@@ -6,19 +6,19 @@ import { useTranslation } from 'react-i18next';
 import { detailsPage, Kebab, navFactory, PageHeading, ResourceLink, SectionHeading, Timestamp } from '../../utils';
 import { SingleExpandableTable } from '../utils/expandable-table';
 import { WebSocketContext } from '../../app';
-import { Button, FormSelect, FormSelectOption, Modal } from '@patternfly/react-core';
+import { Button } from '@patternfly/react-core';
 import { Link } from 'react-router-dom';
 import '../utils/help.scss';
 import * as _ from 'lodash';
 import { FilterIcon } from '@patternfly/react-icons';
 import { DropdownToggle, DropdownToggleCheckbox } from '@patternfly/react-core';
-import { Dropdown, DropdownItem, KebabToggle } from '@patternfly/react-core';
-import { FileUpload } from '@patternfly/react-core';
-import FileUploadIcon from '@patternfly/react-icons/dist/esm/icons/file-upload-icon';
-import { Form, FormGroup, TextInput } from '@patternfly/react-core';
+import { Dropdown } from '@patternfly/react-core';
 import { NavBar } from '../../utils';
 import { Table, TableHeader, TableBody, TableProps } from '@patternfly/react-table';
 import { APP_SOCKET, SERVICE_SOCKET } from './sas-string';
+import { SasAppKebab } from './sas-kebab';
+import { AddVersionModal, AppDeleteModal, AppDeployModal, AppStopModal, ReplicaModal, VersionDeleteModal, VersionSelectModal } from './sas-modal';
+import { InlineEditCell, ModalPage } from './sas-utils';
 
 const kind = 'SasApp';
 
@@ -52,93 +52,6 @@ export const InnerTable = ({ innerDatas, data }) => {
       </div>
     </>
   );
-};
-
-const SasKebab = ({ status, handleModalToggle, data }) => {
-  const [isOpen, setIsOpen] = React.useState(false);
-  const onToggle = (isOpen: boolean) => {
-    setIsOpen(isOpen);
-  };
-
-  const onFocus = () => {
-    const element = document.getElementById('toggle-kebab');
-    element.focus();
-  };
-
-  const onSelect = () => {
-    setIsOpen(false);
-    onFocus();
-  };
-
-  const dropdownItemsReady = [
-    <DropdownItem
-      key="add-version"
-      component="button"
-      onClick={() => {
-        handleModalToggle(data, 'addversion');
-      }}
-    >
-      버전 추가
-    </DropdownItem>,
-    <DropdownItem
-      key="app-deploy"
-      component="button"
-      onClick={() => {
-        handleModalToggle(data, 'appdeploy');
-      }}
-    >
-      앱 배포
-    </DropdownItem>,
-    <DropdownItem
-      key="app-delete"
-      component="button"
-      onClick={() => {
-        handleModalToggle(data, 'appdelete');
-      }}
-    >
-      앱 삭제
-    </DropdownItem>,
-  ];
-  const dropdownItemsRunnung = [
-    <DropdownItem
-      key="addversion"
-      component="button"
-      onClick={() => {
-        handleModalToggle(data, 'addversion');
-      }}
-    >
-      버전 추가
-    </DropdownItem>,
-    <DropdownItem
-      key="version-select"
-      component="button"
-      onClick={() => {
-        handleModalToggle(data, 'versionselect');
-      }}
-    >
-      버전 선택
-    </DropdownItem>,
-    <DropdownItem
-      key="replica"
-      component="button"
-      onClick={() => {
-        handleModalToggle(data, 'replica');
-      }}
-    >
-      레플리카 수 수정
-    </DropdownItem>,
-    <DropdownItem
-      key="app-stop"
-      component="button"
-      onClick={() => {
-        handleModalToggle(data, 'appstop');
-      }}
-    >
-      앱 중지
-    </DropdownItem>,
-  ];
-  const dropdownItems = status === 'Archived' ? dropdownItemsReady : dropdownItemsRunnung;
-  return <Dropdown className="my-dropdown" onSelect={onSelect} toggle={<KebabToggle className="sas-kebab-min" onToggle={onToggle} />} isOpen={isOpen} isPlain dropdownItems={dropdownItems} position={'right'} />;
 };
 
 const SasAppTable = props => {
@@ -220,7 +133,7 @@ const SasAppTable = props => {
       },
       {
         className: Kebab.columnClass,
-        title: <SasKebab status={obj.STATUS} handleModalToggle={props.handleModalToggle} data={obj} />,
+        title: <SasAppKebab status={obj.STATUS} handleModalToggle={props.handleModalToggle} data={obj} />,
       },
     ];
   };
@@ -231,222 +144,6 @@ const SasAppTable = props => {
     return <InnerTable key="InnerTable" innerDatas={data.VERSIONS} data={data.VERSIONS} />;
   };
   return <SingleExpandableTable header={SasAppColumns} itemList={SasAppList} rowRenderer={rowRenderer} innerRenderer={innerRenderer} compoundParent={3} />;
-};
-export const ModalPage = ({ isModalOpen, handleModalToggle, titleModal, InnerPage, submit }) => {
-  const actions = [
-    <Button key="cancel" variant="primary" onClick={handleModalToggle}>
-      취소
-    </Button>,
-    <Button key="confirm" variant="secondary" onClick={submit}>
-      {titleModal[1]}
-    </Button>,
-  ];
-  return (
-    <Modal width={600} title={titleModal[0]} isOpen={isModalOpen} onClose={handleModalToggle} actions={actions}>
-      {InnerPage}
-    </Modal>
-  );
-};
-
-const AddVersionModal = ({ appName, setSubmitData }) => {
-  const [value, setValue] = React.useState(null);
-  const [filename, setFilename] = React.useState('');
-  const [version, setVersion] = React.useState('');
-  const [description, setDescription] = React.useState('');
-  React.useEffect(() => {
-    setSubmitData({ modal: 'AddVersionModal', value, filename, version, description });
-  }, [value, filename, version, description]);
-  const handleUpload = files => {
-    const file = files[0];
-    setValue(file);
-    setFilename(file.name);
-  };
-  const handleVersionChange = (name: string) => {
-    setVersion(name);
-  };
-
-  const handleDescriptionChange = (email: string) => {
-    setDescription(email);
-  };
-
-  return (
-    <>
-      <Form isHorizontal={false}>
-        <div>앱 이름</div>
-        <div>{appName}</div>
-        <FormGroup label="jar 파일" isRequired fieldId="simple-form-jar-02">
-          <FileUpload clearButtonText={''} id="customized-preview-file" value={value} filename={filename} filenamePlaceholder="Drag and drop a file or upload one" onLoad={handleUpload} hideDefaultPreview browseButtonText="검색...">
-            {value && (
-              <div className="pf-u-m-md">
-                <FileUploadIcon size="lg" /> Custom preview here for your {value.size}-byte file named {value.name}
-              </div>
-            )}
-          </FileUpload>
-        </FormGroup>
-        <FormGroup label="버전" isRequired fieldId="simple-form-version-03">
-          <TextInput value={version} isRequired type="text" id="horizontal-form-name" aria-describedby="horizontal-form-name-helper" name="horizontal-form-name" onChange={handleVersionChange} />
-        </FormGroup>
-        <FormGroup label="설명" fieldId="simple-form-description-02">
-          <TextInput value={description} isRequired type="text" id="horizontal-form-name" aria-describedby="horizontal-form-name-helper" name="horizontal-form-name" onChange={handleDescriptionChange} />
-        </FormGroup>
-      </Form>
-    </>
-  );
-};
-const AppDeployModal = ({ appData, setSubmitData }) => {
-  const [version, setVersion] = React.useState('');
-  const [description, setDescription] = React.useState('');
-  const [replica, setReplica] = React.useState('');
-  const [target, setTarget] = React.useState('');
-  const versionOptions = [];
-  appData.VERSIONS.map(version => {
-    versionOptions.push({ value: version.DESCRIPTION, label: version.VERSION, disabled: false });
-  });
-  React.useEffect(() => {
-    setDescription(appData?.VERSIONS[0]?.DESCRIPTION);
-  }, []);
-  const targetOptions = [{ value: appData.POOL_ID, label: appData.POOL_ID, disabled: false }];
-
-  React.useEffect(() => {
-    setSubmitData({ modal: 'AppDeployModal', version, description, replica, target });
-  }, [version, description, replica, target]);
-  const handleReplicaChange = (name: string) => {
-    setReplica(name);
-  };
-
-  const handleTargetChange = (value: string, _event: React.FormEvent<HTMLSelectElement>) => {
-    setTarget(value);
-  };
-
-  const handleOptionChange = (value: string, _event: React.FormEvent<HTMLSelectElement>) => {
-    setVersion(value);
-    setDescription(value);
-  };
-
-  return (
-    <>
-      <Form isHorizontal={false}>
-        <div>앱 이름</div>
-        <div>{appData.APP_NAME}</div>
-        <FormGroup label="버전" fieldId="simple-form-version-03">
-          <FormSelect value={version} onChange={handleOptionChange} id="horizontal-form-title" name="horizontal-form-title" aria-label="Your title">
-            {versionOptions.map((option, index) => (
-              <FormSelectOption isDisabled={option.disabled} key={index} value={option.value} label={option.label} />
-            ))}
-          </FormSelect>
-        </FormGroup>
-        <FormGroup label="설명" fieldId="simple-form-description-02">
-          <div>{description}</div>
-        </FormGroup>
-        <FormGroup label="레플리카 수" isRequired fieldId="simple-form-version-03">
-          <TextInput value={replica} isRequired type="text" id="horizontal-form-name" aria-describedby="horizontal-form-name-helper" name="horizontal-form-name" onChange={handleReplicaChange} />
-        </FormGroup>
-        <FormGroup label="타겟 워커 노드 풀" fieldId="simple-form-version-03">
-          <FormSelect value={target} onChange={handleTargetChange} id="horizontal-form-title" name="horizontal-form-title" aria-label="Your title">
-            {targetOptions.map((option, index) => (
-              <FormSelectOption isDisabled={option.disabled} key={index} value={option.value} label={option.label} />
-            ))}
-          </FormSelect>
-        </FormGroup>
-      </Form>
-    </>
-  );
-};
-const AppDeleteModal = ({ appData, setSubmitData }) => {
-  React.useEffect(() => {
-    setSubmitData({ modal: 'AppDeleteModal' });
-  }, []);
-
-  return (
-    <>
-      <Form isHorizontal={false}>
-        <div>앱 삭제</div>
-        <div>{`${appData.APP_NAME} 앱을 삭제하시겠습니까?`}</div>
-      </Form>
-    </>
-  );
-};
-const VersionSelectModal = ({ appData, setSubmitData }) => {
-  const [version, setVersion] = React.useState('');
-  const [description, setDescription] = React.useState('');
-  const versionOptions = [];
-  appData.VERSIONS.map(version => {
-    versionOptions.push({ value: version.DESCRIPTION, label: version.VERSION, disabled: false });
-  });
-  React.useEffect(() => {
-    setDescription(appData?.VERSIONS[0]?.DESCRIPTION);
-  }, []);
-
-  React.useEffect(() => {
-    setSubmitData({ modal: 'VersionSelectModal', version, description });
-  }, [version, description]);
-
-  const handleOptionChange = (value: string, _event: React.FormEvent<HTMLSelectElement>) => {
-    setVersion(value);
-    setDescription(value);
-  };
-
-  return (
-    <>
-      <Form isHorizontal={false}>
-        <div>앱 이름</div>
-        <div>{appData.APP_NAME}</div>
-        <FormGroup label="버전" fieldId="simple-form-version-03">
-          <FormSelect value={version} onChange={handleOptionChange} id="horizontal-form-title" name="horizontal-form-title" aria-label="Your title">
-            {versionOptions.map((option, index) => (
-              <FormSelectOption isDisabled={option.disabled} key={index} value={option.value} label={option.label} />
-            ))}
-          </FormSelect>
-        </FormGroup>
-        <FormGroup label="설명" fieldId="simple-form-description-02">
-          <div>{description}</div>
-        </FormGroup>
-      </Form>
-    </>
-  );
-};
-const ReplicaModal = ({ appData, setSubmitData }) => {
-  const [inputValue, setInputValue] = React.useState(parseInt(appData.REPLICAS));
-
-  React.useEffect(() => {
-    setSubmitData({ modal: 'ReplicaModal', inputValue });
-  }, [inputValue]);
-
-  const handleInputChange = event => {
-    let newValue = event.target.value;
-    newValue = parseInt(newValue);
-    setInputValue(newValue);
-  };
-
-  const handleIncrement = () => {
-    setInputValue(inputValue + 1);
-  };
-
-  const handleDecrement = () => {
-    setInputValue(inputValue - 1);
-  };
-
-  return (
-    <div>
-      <button onClick={handleIncrement}>+</button>
-      <input type="number" value={inputValue} onChange={handleInputChange} />
-      <button onClick={handleDecrement}>-</button>
-    </div>
-  );
-};
-const AppStopModal = ({ appData, setSubmitData }) => {
-  React.useEffect(() => {
-    setSubmitData({ modal: 'AppStopModal' });
-  }, []);
-
-  return (
-    <>
-      <Form isHorizontal={false}>
-        <div>앱 중지</div>
-        <div>{`${appData.APP_NAME} 앱의 배포를 중지하시겠습니까?`}</div>
-      </Form>
-    </>
-  );
 };
 
 export const SasAppPage = () => {
@@ -538,10 +235,10 @@ export const SasAppPage = () => {
       setData(newData);
     }
   };
-
   const toggle1Checkbox = checkboxes[0];
   const toggle2Checkbox = checkboxes[1];
   const toggle3Checkbox = checkboxes[2];
+
   return (
     <>
       <Helmet>
@@ -604,6 +301,31 @@ export const SasAppsDetailsPage = props => {
   const [data, setData] = React.useState([]);
   const [singleData, setSingleData] = React.useState(null);
   const [servicedata, setServiceData] = React.useState(null);
+  const [isModalOpen, setIsModalOpen] = React.useState(false);
+  const [titleModal, setTitleModal] = React.useState(['', '']);
+  const [InnerPage, setInnerPage] = React.useState(<></>);
+  const [submitData, setSubmitData] = React.useState({});
+  const submit = () => {
+    console.log(123, submitData);
+    setIsModalOpen(!isModalOpen);
+  };
+  const handleModalToggle = (selectedData, type) => {
+    switch (type) {
+      case 'versiondelete':
+        setTitleModal(['버전 삭제', '삭제']);
+        setInnerPage(<VersionDeleteModal appData={selectedData} setSubmitData={setSubmitData} />);
+        break;
+      case 'appdeploy':
+        setTitleModal(['앱 배포', '배포']);
+        setInnerPage(<AppDeployModal appData={selectedData} setSubmitData={setSubmitData} />);
+        break;
+      default:
+        break;
+    }
+
+    setIsModalOpen(!isModalOpen);
+  };
+
   React.useEffect(() => {
     webSocket.ws?.send(APP_SOCKET);
     webSocket.ws?.send(SERVICE_SOCKET);
@@ -627,16 +349,28 @@ export const SasAppsDetailsPage = props => {
       }
     });
   }, [data, name]);
-
+  const handleUpdateDescription = (index, newDescription) => {
+    const updatedRows = [...singleData.VERSIONS];
+    updatedRows[index].DESCRIPTION = newDescription;
+    setSingleData({ ...singleData, VERSIONS: updatedRows });
+  };
   const columns: TableProps['cells'] = ['버전', '설명', 'Jar 파일명', '생서일시', ''];
-  const rows: TableProps['rows'] = singleData?.VERSIONS.map(repo => [repo.VERSION, repo.DESCRIPTION, repo.JAR_NAME, repo.CREATED_AT, '']);
   const serviceColumns: TableProps['cells'] = ['이름', '앱', '크론 수', ''];
   const serviceRows: TableProps['rows'] = servicedata
     ?.filter(repo => {
       if (repo.APP_NAME === appName) return true;
       return false;
     })
-    .map(repo => ({ cells: [repo.SERVICE_PACKAGE, repo.APP_NAME, repo.CRON, ''] }));
+    .map(repo => ({
+      cells: [
+        repo.SERVICE_PACKAGE,
+        repo.APP_NAME,
+        repo.CRON,
+        <td className="td-in-kebab">
+          <SasAppKebab status={'Service'} handleModalToggle={handleModalToggle} data={repo} />
+        </td>,
+      ],
+    }));
   return (
     <div>
       <PageHeading detail={true} title={appName} badge={props.badge} icon={props.icon}></PageHeading>
@@ -667,10 +401,28 @@ export const SasAppsDetailsPage = props => {
       <div className="co-m-pane__body">
         <SectionHeading text={'버전'} />
         {singleData && (
-          <Table aria-label="Simple Table" cells={columns} rows={rows}>
-            <TableHeader />
-            <TableBody />
-          </Table>
+          <table className="table table-bordered table-hover">
+            <thead>
+              <tr>
+                {columns.map((col, index) => (
+                  <th key={index}>{col}</th>
+                ))}
+              </tr>
+            </thead>
+            <tbody>
+              {singleData.VERSIONS.map((row, rowIndex) => (
+                <tr key={rowIndex}>
+                  <td>{row.VERSION}</td>
+                  <InlineEditCell value={row.DESCRIPTION} onUpdate={newDescription => handleUpdateDescription(rowIndex, newDescription)} />
+                  <td>{row.JAR_NAME}</td>
+                  <td>{row.CREATED_AT}</td>
+                  <td className="td-in-kebab">
+                    <SasAppKebab status={'Version'} handleModalToggle={handleModalToggle} data={row} />
+                  </td>
+                </tr>
+              ))}
+            </tbody>
+          </table>
         )}
       </div>
       <div className="co-m-pane__body">
@@ -682,6 +434,7 @@ export const SasAppsDetailsPage = props => {
           </Table>
         )}
       </div>
+      <ModalPage isModalOpen={isModalOpen} handleModalToggle={handleModalToggle} InnerPage={InnerPage} titleModal={titleModal} submit={submit}></ModalPage>
     </div>
   );
 };
